ECOPROJECT-3416 | feat: Baremetal cost prediction
This PR add a cost estimation for baremetal self managed OCP/OVE.

The calculation is based on:
https://www.redhat.com/en/resources/self-managed-openshift-subscription-guide

We predefine vCPU for calc based on vCenter VMs. We let user define the Node HW,
we can pre-define it, but it's hard as each node could be different.

We also offer user to define OCP/OVE or SLA.

Signed-off-by: Ondra Machacek <omachace@redhat.com>